third_party/blink/web_tests/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/transitions-002-expected.txt

This is a testharness.js-based test.
FAIL Scrollbars do not cause a transition of background-color assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 64, 128)"
Harness: the test ran to completion.
